requirements.txt
Flask
uuid
PyJWT
torch
numpy
torchvision
matplotlib
pandas
dask[complete]
tensorboard
zarr
fsspec
requests
aiohttp
seaborn
scikit-learn
lightning